+++ src/View/View/AdminView.php
@@ -22,6 +22,7 @@
  */
 namespace MeCms\View\View;
 
+use App\View\AppView as BaseView;
 use MeCms\View\View\AppView;
 
 /**
@@ -31,12 +32,20 @@ class AdminView extends AppView {
 	/**
      * Initialization hook method
 	 * @see http://api.cakephp.org/3.1/class-Cake.View.View.html#_initialize
-	 * @uses MeCms\View\View::initialize()
+	 * @uses App\View\AppView::initialize()
 	 */
     public function initialize() {
-		parent::initialize();
+		BaseView::initialize();
 		
 		//Loads helpers
+		$this->loadHelper('Html', ['className' => 'MeTools.Html']);
+		$this->loadHelper('MeTools.Dropdown');
+		$this->loadHelper('MeTools.Form');
+		$this->loadHelper('MeTools.Asset');
+		$this->loadHelper('MeTools.Library');
+		$this->loadHelper('MeTools.Thumb');
+		$this->loadHelper('MeTools.Paginator');
+		$this->loadHelper('MeCms.Auth');
 		$this->loadHelper('Menu', ['className' => 'MeCms.MenuBuilder']);
 	}
 	
@@ -46,8 +55,7 @@ class AdminView extends AppView {
 	 * @param string|NULL $layout Layout to use
 	 * @return string|NULL Rendered content or NULL if content already rendered and returned earlier
 	 * @see http://api.cakephp.org/3.1/class-Cake.View.View.html#_render
-     * @throws Cake\Core\Exception\Exception
-	 * @uses MeCms\View\View\AppView::render()
+	 * @uses App\View\AppView::render()
 	 * @uses layout
 	 * @uses viewVars
 	 */
@@ -65,6 +73,21 @@ class AdminView extends AppView {
 			'5' => sprintf('5 - %s', __d('me_cms', 'Very high'))
 		];
 		
-		return parent::render($view, $layout);
+		return BaseView::render($view, $layout);
+	}
+	
+	/**
+	 * Renders a layout. Returns output from _render(). Returns false on error. Several variables are created for use in layout
+	 * @param string $content Content to render in a view, wrapped by the surrounding layout
+	 * @param string|null $layout Layout name
+	 * @return mixed Rendered output, or false on error
+	 * @see http://api.cakephp.org/3.1/source-class-Cake.View.View.html#477-513
+	 * @uses _getTitleForLayout()
+	 */
+	public function renderLayout($content, $layout = NULL) {
+		//Assigns the title for layout
+		$this->assign('title', $this->_getTitleForLayout());
+				
+		return BaseView::renderLayout($content, $layout);
 	}
 }